llm/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
faiss-cpu==1.9.0
sentence-transformers>=2.7.0
ollama==0.1.6
pydantic==2.5.0
python-multipart==0.0.6
gTTS>=2.5.0
